MyInfo component inserted in an external file and imported to index.js
