+++ project_1/src/MyInfo.js
@@ -1,0 +1,22 @@
+import React from "react"
+
+
+function MyInfo()
+{
+    return (
+        <div>
+            <h1>Willian A. Lopes</h1>
+            <p>Hi, I'm learning React! xD</p>
+            <ol>
+                {/* This is my top three vacation places: */}
+                <li>Egipt</li>
+                <li>Russia</li>
+                <li>Peru</li>
+            </ol>
+            <a href="http://www.google.com">Google</a>
+        </div>
+    )
+}
+
+
+export default MyInfo;
