+++ project_1/src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
+import MyInfo from './MyInfo'
 import './index.css';
 
 function MyApplication()
@@ -14,22 +15,6 @@ function MyApplication()
     )
 }
 
-
-function MyInfo()
-{
-    return (
-        <div>
-            <h1>Willian A. Lopes</h1>
-            <p>Hi, I'm learning React! xD</p>
-            <ol>
-                This is my top three vacation places:
-                <li>Egipt</li>
-                <li>Russia</li>
-                <li>Peru</li>
-            </ol>
-        </div>
-    )
-}
